+++ QA Tests/C# Tests/UNOPS.PAO.Business.Tests/Authorization/RolePermissionComprehensiveTests.cs
@@ -377,7 +377,10 @@ namespace UNOPS.PAO.Business.Tests.Authorization
 
         public void Dispose()
         {
-            _context.Database.EnsureDeleted();
+            if (TestEnvironment.UseInMemory)
+            {
+                _context.Database.EnsureDeleted();
+            }
             _context.Dispose();
         }
     }
