+++ QA Tests/C# Tests/UNOPS.PAO.Business.Tests/Managers/DocumentTypeManagerTests.cs
@@ -488,7 +488,10 @@ namespace UNOPS.PAO.Business.Tests.Managers
 
         public void Dispose()
         {
-            _context.Database.EnsureDeleted();
+            if (TestEnvironment.UseInMemory)
+            {
+                _context.Database.EnsureDeleted();
+            }
             _context.Dispose();
         }
     }
